llm_analysis_service/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-pubsub
openai